Se agregan cambios del grupo

+++ src/ejercicio3/Edificio.java
@@ -1,5 +1,5 @@
 package ejercicio3;
 
 public interface Edificio {
-	int getSuperficieEdificio();
+	double getSuperficieEdificio();
 }
